Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/html/semantics/scripting-1/the-script-element/module/choice-of-error-3-expected.txt

CONSOLE MESSAGE: line 2: [object Object]

FAIL Evaluation errors are cached in intermediate module scripts assert_equals: expected 5 but got 4
